debugs block and detect

+++ BUPT_TowerDefence.py
@@ -322,13 +322,13 @@ class Level_choose(cocos.menu.Menu):
 
         self.addable = False
         self.auto_new = 0
-        global scroller
+        global scroller,speed_1,block_1,block_1_R
+        speed_1 = 1
         scroller = cocos.layer.ScrollingManager()
         self.e_list = []
 
         self.skill = [[True,0],[True,0]]            #两个数组第一个是技能1 每一个有两个参数：参数1：现在是否可以释放 参数2：冷却时间计数
-        global speed_1
-        speed_1 = 1
+
         self.player_damage = 5
 
         self.eb_count=0
@@ -337,7 +337,7 @@ class Level_choose(cocos.menu.Menu):
         self.t_list=[]
         self.t_list.append(Teammate(animation.myE_skeleton.skeleton,animation.myE_skin.skin,"/animation/q.anim","/animation/E_attack.anim","/animation/frozen.anim",0))
         # self.player_2 = Player_2()
-        self.e_list.append([Enemy_1(animation.test_skeleton.skeleton,animation.test_skin.skin,"/animation/2t.anim","/animation/E_attack.anim","/animation/frozen.anim",0), False, 0])   #第一个是对象 第二个是是否死亡 第三个是count（子弹击中）
+        self.e_list.append([Enemy_1(animation.test_skeleton.skeleton,animation.test_skin.skin,"/animation/2t.anim","/animation/E_attack.anim","/animation/frozen.anim",0), False])   #第一个是对象 第二个是是否死亡 第三个是count（子弹击中）
         self.enemy_1_dead = False
         self.add_e_count=0
         self.coll_manager = cm.CollisionManagerBruteForce()
@@ -384,7 +384,7 @@ class Level_choose(cocos.menu.Menu):
                 self.skill_1.color = (125,125,125)
                 self.skill[0][0] = False
                 self.skill[0][1] = 0
-                self.player_damage = 50
+                self.player_damage = 10
         else:
             if (self.skill[0][1] <= 200):
                 self.skill[0][1] += 1
@@ -425,7 +425,7 @@ class Level_choose(cocos.menu.Menu):
 
     def add_enemy(self):
         if self.addable==False:
-            self.auto_new+=1
+            self.auto_new += 1
         if self.auto_new>=400:
             self.addable=True
             self.auto_new=0
@@ -451,66 +451,65 @@ class Level_choose(cocos.menu.Menu):
             if self.add_e_count>40:
                 self.add_e_count=0
 
+    def death_detect(self):
+        global scroller
+        if self.player_1.life <= 0:
+            bg_1 = BG(bg_name="img/fail_bg.png")  # 1.获取背景图片路径
+            self.game_menu = Game_menu_f()
+            scene_2 = cocos.scene.Scene(bg_1,Fail_Layer(),Drag(),MouseDisplay(), self.game_menu)
+            director.replace(scenes.transitions.SlideInBTransition(scene_2, duration=1))
+        for enemy in self.e_list:
+            if enemy[0].life <= 0:
+                enemy[1] = True
+                scroller.remove(enemy[0])
+                self.scene_3.unschedule(enemy[0].update_position)
+                self.scene_3.unschedule(enemy[0].status_detect)
+                self.e_list.remove(enemy)
+        for teammate in self.t_list:
+            if teammate.life <= 0:
+                scroller.remove(teammate)
+                self.scene_3.unschedule(teammate.update_position)
+                self.scene_3.unschedule(teammate.status_detect)
+                self.t_list.remove(teammate)
+
+
     def enemy_detect(self):
+        self.death_detect()
+        global block_1,block_1_R,block
         for enemy in self.e_list:
             if not enemy[1]:           #没死的话
                 if self.coll_manager.they_collide(self.player_1, enemy[0]):     #主角和enemy检测
-                    self.player_1.mover.block_R = True
-                    if enemy[0].status == 1:
+                    block_1_R = True
+                    if enemy[0].status == 1 or enemy[0].status == 3:        #enemy可以进行攻击
                         enemy[0].Attack = True
-                    if enemy[0].status == 5:
-                        self.player_1.mover.block = True
-                        if self.player_1.life < 2:
-                            self.player_1.life = 0
-                        else:
-                            self.player_1.life -= 1
-                            self.player_1.beheat = True
+                        self.player_1.beheat = True
+                    if enemy[0].status == 4:        #enemy攻击态
+                        block_1 = True
+                        self.player_1.life -= 1
+                        self.player_1.beheat = True
                     else:
-                        self.player_1.mover.block = False
+                        block_1 = False
                 else:
-                    self.player_1.mover.block_R = False
-                    self.player_1.mover.block = False
+                    block_1_R = False
 
-                for teammate in self.t_list:                                #主角和小弟检测
-                    if self.coll_manager.they_collide(teammate, enemy[0]):
-                        teammate.mover.block_R = True
-                        if enemy[0].status == 1:
+                for teammate in self.t_list:                                #敌人和小弟检测
+                    if self.coll_manager.they_collide(teammate, enemy[0]):      #得random或者怎样 调整怪物先手顺序 或者用speed判断
+                        if enemy[0].status == 1 or enemy[0].status == 3:
                             enemy[0].Attack = True
-                        if enemy[0].status == 5:
-                            teammate.mover.block = True
-                            if teammate.life < 2:
-                                teammate.life = 0
-                                scroller.remove(teammate)
-                                self.scene_3.unschedule(teammate.update_position)
-                                self.scene_3.unschedule(teammate.status_detect)
-                                self.t_list.remove(teammate)
-                            else:
-                                teammate.life -= 1
-                                teammate.beheat = True
-                        else:
-                            teammate.mover.block = False
-                if enemy[2] != 0:
-                    enemy[2] += 1
-                if enemy[2] >= 7:
-                    enemy[2] = 0
-                if self.coll_manager.they_collide(self.player_1.bullet, enemy[0]):
-                    if enemy[2] == 0:
-                        enemy[2] += 1
-                        self.player_1.refresh()
-                        if enemy[0].life <= 20:
-                            enemy[0].life = 0
-                            enemy[1] = True
-                            scroller.remove(enemy[0])
-                            # self.coll_manager.remove_tricky(enemy[0])  #加了 好像也remove不掉 还无法切场景
-                            self.scene_3.unschedule(enemy[0].update_position)
-                            self.scene_3.unschedule(enemy[0].status_detect)
-                            self.e_list.remove(enemy)
-                        else:
-                            enemy[0].life = enemy[0].life - self.player_damage
-                            enemy[0].beheat = True
-            else:
-                self.player_1.mover.block_R = False
-                self.player_1.mover.block = False
+                            teammate.beheat = True
+                        elif enemy[0].status == 5:
+                            enemy[0].life -= 2   
+                        if teammate.status == 1:
+                            teammate.Attack = True
+                            enemy[0].beheat = True 
+                        elif teammate.status == 5:
+                            teammate.life -= 2
+
+                if self.coll_manager.they_collide(self.player_1.bullet, enemy[0]):  #子弹和敌人检测
+                    enemy[0].beheat = True
+                    enemy[0].life = enemy[0].life - self.player_damage
+                    self.player_1.refresh()
+
 
     def update(self, dt):
         self.skill_detect()
@@ -518,7 +517,6 @@ class Level_choose(cocos.menu.Menu):
         self.ebase_detect()
         self.enemy_detect()
 
-        
         if self.m_layer.enemy_base.dead:
             self.m_layer.enemy_base.dead = False
             bg_1 = BG(bg_name="img/fail_bg.png")  # 1.获取背景图片路径
@@ -574,19 +572,15 @@ class draw_rec(cocos.layer.util_layers.ColorLayer):
 class Mover_1(cocos.actions.BoundedMove):
     def __init__(self):
         super().__init__(2300, 1430)  # it should be bigger than the size of the picture
-        self.block = False
-        self.block_R = False
-
+        global block_1,block_1_R,speed_1,scroller
     def step(self, dt):  # add block
-        if not self.block:
+        if not block_1:
             super().step(dt)
-            global speed_1
             vel_x = (keyboard[key.D] - keyboard[key.A]) * 400 * speed_1
-            if self.block_R and vel_x > 0:
+            if block_1_R and vel_x > 0:
                 vel_x = 0
             vel_y = 0
             self.target.velocity = (vel_x, vel_y)
-            global scroller
             scroller.set_focus(self.target.x, self.target.y)
 
 
@@ -607,9 +601,8 @@ class Mover_2(cocos.actions.BoundedMove):
 class Mover_3(cocos.actions.BoundedMove):
     def __init__(self,num):
         super().__init__(2300, 1430)  # it should be bigger than the size of the picture
-        self.block = False
-        self.num = num
         global block
+        self.num = num
     def step(self, dt):  # add block
         if not block[self.num]:
             super().step(dt)
@@ -727,7 +720,7 @@ class Player_1(cocos.layer.ScrollableLayer):
                 self.skin.remove_action(self.skin.actions[0])
 
     def update_position(self, dt):
-        if not self.mover.block:
+        if not block_1:
             self.skin.position = self.spr.position  # !!!!!!! self.position = -(self.skin.position-600)
             x, y = self.skin.position
             self.life_bar.position = (x, y + 160)
@@ -743,13 +736,15 @@ class Player_1(cocos.layer.ScrollableLayer):
     def refresh(self):
         self.count = 0
         self.bullet.position = -100,-100
-        self.mover.block = False
+        global block_1
+        block_1 = False
         self.bullet_move = False
 
     def recover(self):
         self.status = 1
         self.remove_all()
-        self.mover.block = False
+        global block_1
+        block_1 = False
 
     def status_detect(self, dt):
         self.bullet.cshape.center = self.bullet.position
@@ -762,9 +757,9 @@ class Player_1(cocos.layer.ScrollableLayer):
                 self.refresh()
         elif self.status == 1:
             if (keyboard[key.J]):
-                self.remove_all()
                 self.status = 4
-                self.mover.block = True
+                global block_1
+                block_1 = True
                 self.skin.do(skeleton.Animate(self.attack))
                 timer = threading.Timer(0.2, self.recover)
                 timer.start()
@@ -774,28 +769,22 @@ class Player_1(cocos.layer.ScrollableLayer):
                 x, y = self.skin.position
                 self.bullet.position = x + 110, y + 70
                 self.bullet_move = True
-            elif (keyboard[key.D] and not self.mover.block_R):  # key right and not attack  得改成状态保持
-                self.remove_all()
-                self.status = 2
-                self.skin.do(cocos.actions.Repeat(skeleton.Animate(self.walk)))
-            elif (keyboard[key.A]):  # key right and not attack
-                self.remove_all()
+            elif ((keyboard[key.D] and not block_1_R)or(keyboard[key.A])):  # key right and not attack  得改成状态保持
                 self.status = 2
                 self.skin.do(cocos.actions.Repeat(skeleton.Animate(self.walk)))
         if(self.status == 2 and not(keyboard[key.D] or keyboard[key.A])):
-                self.remove_all()
-                self.status = 1
-        if self.beheat:
+                self.recover()
+        if self.beheat and self.status != 3:
             self.remove_all()
             self.status = 3
             self.beheat = False
-            self.mover.block = True
+            block_1 = True
             self.skin.do(skeleton.Animate(self.frozen))
-            timer = threading.Timer(0.7, self.recover)
+            timer = threading.Timer(0.5, self.recover)
             timer.start()
                 
         if self.life<=0:
-            del self
+            del self        #这里确定没有问题？？
             bg_1 = BG(bg_name="img/fail_bg.png")  # 1.获取背景图片路径
             game_menu = Game_menu_f()
             scene_2 = cocos.scene.Scene(bg_1,Fail_Layer(),Drag(),MouseDisplay(), game_menu)
@@ -1025,6 +1014,8 @@ if __name__ == '__main__':
     # 全局变量
     target_x, target_y = (0, 0)
     block = []
+    block_1 = False
+    block_1_R = False
     block_2 = False
     img_name = ""
     pygame.mixer.init()
